build and run testngpp tests on mingw success.

+++ tests/3rdparty/testngpp/tests/3rdparty/mockcpp/src/JmpCodeX86.h
@@ -20,7 +20,7 @@
 const unsigned char jmpCodeTemplate[]  = { 0xE9, 0x00, 0x00, 0x00, 0x00 };
 #define SET_JMP_CODE(base, from, to) do { \
         *(unsigned long*)(base + 1) = \
-            (unsigned long)to - (unsigned long)from - sizeof(jmpCodeTemplate); \
+            (unsigned long long)to - (unsigned long long)from - sizeof(jmpCodeTemplate); \
    } while(0)
 
 #endif
